refactor(frontend): remove duplicated code
